+++ docs/src/monacoEditor.d.ts
