+++ docs/src/monacoEditor.js
